+++ models/demos/t3000/llama2_70b/tests/test_llama_device_perf.py
@@ -5,6 +5,7 @@
 import sys
 
 import pytest
+from tracy.process_model_log import get_latest_ops_log_filename, run_device_profiler
 
 from models.demos.t3000.llama2_70b.tests.test_llama_model import (
     DEVICE_PERF_START_SIGNPOST,
@@ -15,7 +16,6 @@ from models.demos.t3000.llama2_70b.tt.llama_common import check_mesh_device, set
 from models.demos.t3000.mixtral8x7b.scripts.op_perf_results import main as calculate_op_perf_results
 from models.perf.device_perf_utils import check_device_perf
 from models.utility_functions import skip_for_grayskull
-from tt_metal.tools.profiler.process_model_log import get_latest_ops_log_filename, run_device_profiler
 
 
 @pytest.mark.parametrize(
